requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.8.1
    # via -r requirements.in
anyio==3.6.1
    # via starlette
asyncpg==0.26.0
    # via databases
click==8.1.3
    # via uvicorn
databases[postgresql]==0.6.1
    # via -r requirements.in
fastapi==0.85.0
    # via -r requirements.in
greenlet==1.1.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
mako==1.2.3
    # via alembic
markupsafe==2.1.1
    # via mako
orjson==3.8.0
    # via -r requirements.in
psycopg2==2.9.4
    # via -r requirements.in
pydantic==1.10.2
    # via fastapi
python-dotenv==0.21.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
sqlalchemy==1.4.41
    # via
    #   alembic
    #   databases
starlette==0.20.4
    # via fastapi
typing-extensions==4.4.0
    # via pydantic
uvicorn==0.18.3
    # via -r requirements.in
